Added view package. Renamed some classes and fixed dice moving when selected.
