--- src/enjug/erijan/games/yatzy/rules/BoxUse.java
@@ -1,8 +1,0 @@
-package enjug.erijan.games.yatzy.rules;
-
-/**
- * Created by Jan Eriksson on 18/11/15.
- */
-public enum BoxUse {
-  DERIVED, TOTAL, SUM
-}
