+++ src/enjug/erijan/games/yatzy/rules/YatzyFactory.java
@@ -2,6 +2,7 @@ package enjug.erijan.games.yatzy.rules;
 
 import enjug.erijan.games.util.DiceHandler;
 import enjug.erijan.games.yatzy.*;
+import enjug.erijan.games.yatzy.view.YatzyGui;
 
 /**
  * Created by Jan Eriksson on 18/11/15.
@@ -9,5 +10,5 @@ import enjug.erijan.games.yatzy.*;
 public interface YatzyFactory {
   ScoreModel getScoreModel(Player player);
   DiceHandler getDice();
-  YatzyGui getGui(YatzyAgentInterface yatzyAgent, DiceHandler diceHandler);
+  YatzyGui getGui(GameControlInterface yatzyAgent, DiceHandler diceHandler);
 }
