+++ src/enjug/erijan/games/yatzy/view/DiceIcons/d6-1.png
