+++ src/enjug/erijan/games/yatzy/view/DiceIcons/d6-2.png
