+++ src/enjug/erijan/games/yatzy/view/DiceIcons/d6-3-bow.png
