+++ src/enjug/erijan/games/yatzy/view/DiceIcons/d6-4-bow.png
