+++ src/enjug/erijan/games/yatzy/view/DiceIcons/d6-5-bow.png
