+++ src/enjug/erijan/games/yatzy/view/DiceIcons/d6-6.png
